Adding initial implementation of dynamic resource provider for Salt
In order to connect salt minions with the salt master we need to pre-create the minion keys. This resource provider uses the Salt API to generate a minion keypair and accept it on the master, so that we can pass the key to the user-data input of an EC2 instance at creation time.