Adding fix for conditional resource imports to avoid loops
In the current implementation the imported resource doesn't get the `pulumi_managed` tag applied so that subsequent runs will try to import the same resource again. This adds a step to use boto3 to apply that tag after generating the ResourceOptions that specify the import but only when the resources are actually being manipulated.