Adding in Pants and starting work to package deploys
In order to be able to support building images via Packer we want to be able to package up the PyInfra deploys so that they can be copied to the target host and executed directly. To make this possible this commit adds the Pants build tool and an initial configuration for packaging the target deployment as a PEX binary.