Renaming module space to be more concise
Removed the `ol` prefix to make it more useful outside of the bounds of Open
Learning. Changed the module name to be "Bilder" which is the german word for "images", as well as being phonetically similar to "builder". This is because the code in this project is used to build servers, as well as having Packer configurations for creating server images.