Adding route management for connection between Micromasters and Operations
Redash runs in the operations VPC and connects to the Micromasters read replica in the deprecated Micromasters VPC. When we run the Pulumi code to update network configurations it inadvertantly removes the route table entry that allows Redash to connect to the MM read replica. This adds a special case to maintain that route so it doesn't regress and cause BI reports to fail for network related reasons.