Adding modified common settings file for edX platform config
The YAML files that are used to configure the LMS and studio applications have a lot of shared boilerplate, including a lot of values that are completely unnecessary or not even used. This adds a consolidated, annotated, and stripped down set of common configuration for both applications to be used as the starting point for generating templates to be used at runtime via consul-template.