Consul-template fixes
Setting the templates as a subdirectory of the consul-template config directory target causes it to try to use the templates as a configuration file which errors out. This moves the configuration directory to be a neighbor of the templates under /etc/consul-template

This also fixes some minor syntax issues, configuration errors, and file permissions.