Adding separate target for Studio to allow for sticky sessions
Studio has some workflows that are stateful, such as importing and exporting courses which requires files to be written and read from the same EC2 instance. This adds separate target groups and ALB listener rules to route requests for studio to a target group with session stickiness enabled so that these stateful workflows don't fail.