Fixing up Concourse worker to prevent constant restarts
The systemd.path unit was being configured to watch too many paths, including a directory that was being constantly modified. This led to a continual restart cycle for the worker nodes. This removes the "clever" approach of watching any path objects in the config and instead explicitly enumerates the paths that we care about watching.