Adding marketing app in list of logout URLs
In order for users to be properly logged out of edx and the frontend application we need to route them to the `/logout` path of the marketing app as well. This adds that to the list of IDAs (Independently Deployed Applications) that edx-platform will route users to for terminating a session.