Fixing compilation of theme during AMI build
The theme wasn't being compiled fully during the AMI build because the configuration necessary to point the update-assets script at the right location isn't getting loaded until the instance is launched. This adds the necessary Ansible values to the playbook file so that the minimal necessary config is in place at build/compilation time.