Splitting edX build into pre- and post-ansible steps to address theme
In order for the theme to build properly it needs to have configuration present in the Ansible parameters so that the YAML has the right values present at compilation time. By passing those values it attempts to run the build during the Ansible process. This moves the step of cloning the repository into a pre-build module so that the theme is present during build to prevent failure in the Ansible stage.