Removing unnecessary permission statements in Postgres role template
Because of having to grant permissions on multiple tables to multiple users it can result in exceeding the row length of system tables in the database. Because we are relying on the app role as the actual operating user then we can remove statements that grant permissions directly to the created user.