Cleaning up and combining log processing for use with Loki
The Loki query language doesn't support working with arrays very well, so it's more useful to have a single attribute to identify the process that is responsible for a set of log lines. This introduces a `log_process` attribute that we can use as a standard format for identifying what is responsible for generating the associated messages. This also moves the Nginx log collection in-line with edxapp logs and ships them to Loki.