Adding wrapper for SOPS to make secrets easier to manage
Pulumi secrets in stacks are opaque and become difficult to manage when they are complex objects. This makes it challenging to understand what the structure and keys are in an object, so that diffs that add or remove attributes are not clear, as well as making it difficult to understand if a required attribute is present in an encrypted value.

There are also cases where secrets need to be shared across stacks, or between Pulumi and PyInfra.

This adds a wrapper for ths SOPS binary to handle decrypting of YAML, JSON, or .env data to be used in those various contexts. It also extracts the secret object for populating edX values into Vault as a first use of this capability.