Adding code to bootstrap newly created Vault clusters
Once we deploy a new Vault cluster into an autoscaling group we need to be able to initialize it and set up the initial set of authentication and audit backends. This adds a script that will initialize the cluster, log the recovery keys, and provision an account to be used by Pulumi for managing the cluster going forward.