Adding automatic setup of Vault provider with userpass auth
Now that we are running new clusters that have the username/password authentication set up with credentials stored in SOPS files we can automate the setup of the Vault provider so that we don't have to do it manually every time. This adds a stack transformation that we can call to set up the provider for any Vault resource that needs one.