Adding pipeline definition for running ocw-to-hugo
In order to manage the cutover from Plone to ocw-studio we need a way to keep the content in sync that is independent of the EC2 build server that is deprecated. This adds a pipeline definition that executes the ocw-to-hugo utility for converting the Plone generated JSON files to the Markdown format expected by Hugo and then uploads that output to the S3 bucket used by OCW Studio to import the legacy content.