Updating consul template usage for edxapp
Up to now we were using a two step rendering process for edx-platform configuration files that relied on pushing a full template to Consul, rendering that template to disk and then running consul-template again to render that updated file. This allows for very fast updates, but is a bit confusing and unwieldy. It was also an optimization to account for edX AMI builds being very slow (~45 minutes).

As part of this work we have split the AMI builds into two stages, with the first being the provisioning of the edx-platform code and the second being much faster, consisting only of adding our own operational customizations. With this change in build/deployment the time to delivery for configuration template changes is short enough that it makes sense to build a new image for those updates. This also means that when we make a change to a template it won't require us to restart the processes on all running instances at the same time which can lead to user-facing timeouts.