Fixing SITE_NAME value for edX deployments to resolve errors with LTI
The `SITE_NAME` value needs to actually be the DNS hostname (e.g. lms.example.com & cms.example.com) for the application that is being interacted with, not an arbitrary string that is shown to end users. This value is required to be set accurately in order for LTI integrations to function properly.