Adding conditional to guard against activating traffic to Mongo Atlas
Consul template is set up to route traffic to Mongo Atlas when the Consul keys are set by the mongodb_atlas project. This adds a conditional to guard against that happening automatically, requiring an opt-in stack setting to populate those values. This will let us deploy the clusters, migrate the data, and then cut over traffic in a second stage when we're ready.