Add management of Vault autopilot config in setup
The autopilot mechanism in Vault allows us to have the cluster self-manage removal of dead nodes automatically. This is necessary to ensure that old instances aren't retained in the expected quorum when deploying new instances. If the old instances are not purged automatically in a short timeframe then they will pollute the voting pool, causing the cluster to enter a broken state when the leader gets cycled out, leaving the cluster without a leader.