Adding fix for compiling proctortrack web worker bundle
Proctortrack relies on a web worker to observe the end-user's browser and the code for that worker is loaded at runtime via the LMS backend. The way to locate the necessary JS bundle is through the `webpack-worker-stats.json` file which is only generated if the proctortrack integration is configured at build time. Because we manage the configuration post-deploy we have to bootstrap the configuration for the web worker at build time. This adds a static definition of the `workers.json` file that webpack relies on to generate the JS bundle and mapping file that lets the worker load the proper file.