Configure course visibility settings for edxapp
For xPro and MITx Online we only want users to be able to access and enroll in courses via the marketing sites. For residential staging we want to prevent students from enrolling in courses that are not live. This sets the course visibility permissions for each of those platforms to be staff only.