Adding a dockerfile for packaging project dependencies
As we move to running the logic in this repo in Concourse pipelines we need a way to provide dependencies to resources for executing PyInfra and Pulumi code. This adds a dockerfile to install the Python dependencies from the Poetry lock file in a virtualenv that can be exposed for use by other stages of a Concourse pipeline.