Explicitly setting PATH for Packer builds to include /usr/local/bin
We are copying the contents of the ol-infrastructure dependency build into the runtime container for the Packer AMI builds in Concourse. This places some binaries in /usr/local/bin which does not appear to resolve in the context of the Packer shell execution. This introduces the `PATH` env var as an explicit value to ensure that we can rule that out as an error.