feat: Add a commit message and pull request template
We have a wide variety of formats for our commits which can often leave out useful
context. This adds a pair of templates to help ensure that we are capturing the needed
information at the time that it is fresh in our minds rather than having to reconstruct
that context in the future.