feat: Create CI pipelines for edxapp AMI builds
In order to reduce the complexity and lead time for deploying changes to edX
installations we need to be able to automatically build new images when inputs to the
build change. This adds new pipeline definitions to handle the AMI builds which can lead
into pipelines for automating the deployment of those images.

This creates pipeline templates and vars for:
- Build a base AMI for a given release of edx-platform
- Trigger a build of our customizations for specific installations of edx-platform when
the base images are updated or our configuration logic changes in GitHub