conf: Add toggle for enabling Mongo cloud backup
In order to add more cost controls for our MongoDB Atlas bill we need to be able to
toggle whether we are running backups on a given deployment. For non-production
environments it is largely unnecessary. This adds a stack config toggle to allow for
turning off backups with the default value being to turn them on.