feat: Pin named releases to specific envs by stack
As we test out new named releases we need to maintain a way for deploying the previous
release to QA and production environments in the event of a security patch or
bugfix. This adds an AMI tag and associated stack configuration to filter the AMI that
is deployed per environment/stack. This lets us build e.g. Nutmeg AMIs to deploy to
residential CI while preventing accidental deployment to QA and production.