fix: Explicitly mark edx-platform repo as safe for git
Git has recently added logic that forces you to use the same user as the owner of the
repo directory due to a potential exploit that will run arbitrary commands. In order to
work around this it is required to configure the repository as safe so that operations
by other users can be performed.