fix: Move edxapp AMI filters to a shared list
The filters for the web and worker AMI lookups are almost identical aside from the name
prefix, yet they were defined twice. This caused an error where an additional filter was
applied to the web image to ensure we are deploying the correct named release during
testing of Nutmeg but the same filter wasn't applied to the worker. This moves the
shared filters to a single list that is then used for both lookups to avoid this error
in the future.