fix: Ensure volume cleanup after Packer builds
By default the `amazon-ebs` builder for Packer doesn't terminate the attached EBS
volumes after the image is created which results in a huge accumulation of old volumes
hanging around and costing a substantial amount. This adds an explicit launch block
device mapping to each Packer template to ensure that we specify that those volumes
should be deleted at the end of the build. In addition it adds a `Name` tag to each
volume and snapshot so they are easier to view in the AWS console.