refactor: Create a single pipeline template for MFEs
We have two dimensions along which we need to manage the MFEs that are deployed for edX;
the different edX installations and their respective environments, and the different
MFEs. This pulls the specific values for each instance of that matrix into separate vars
files so that we can keep a single pipeline definition as a template for the different
pipeline instances that we need.