fix: Add peer connection between data and mitx-staging VPCs
We recently added logic to grant access from the Data VPC to various RDS
instances (including edxapp) for data integration purposes via security group
references. In order for those references to work across VPCs they need to be
peered. The only peering connection that was missing was from data -> mitx-staging and
this resolves that situation.