fix: Allow Airbyte version to be updated
The Airbyte version was accidentally hardcoded to a static value, preventing us from
being able to build a new image with a new version. This adds a pass-through of the
Airbyte version to the Consul template via the consul-template environment. It also adds
a github-release trigger to the AMI build pipeline so that new versions will be built as
they are released.