feat: Add Pydantic wrapper for Concourse pipelines
This adds a set of Pydantic models that encode the structure of Concourse pipelines,
allowing us to easily create pipeline definitions in pure Python. This also adds a
definition for the MFE pipelines and a meta pipeline to generate the various
combinations of MFEs that we require across our Open edX deployments.