fix: Explicitly set character set in Django DB connection
In the Nutmeg and later releases the course outline is stored in MySQL/MariaDB which has
led to errors with courses that use unicode characters longer than 3 bytes. The
underlying DB has already been converted to use `utf8mb4` character set and collation,
but usage of e.g. emoji characters is still erroring out. This adds the character set
configuration in the Django connection settings as well.

https://discuss.openedx.org/t/nutmeg-upgrade-outlines-py-openedx-core-djangoapps-content-learning-sequences-models-learningcontext-doesnotexist-learningcontext-matching-query-does-not-exist/