fix: Increase Fastly timeout for residential due to slow APIs
There are a handful of API endpoints in Open edX that are slow to respond. This
increases the timeout for the edx-platform backend in Fastly to allow for those
endpoints to return without timing out because there are legitimate cases where we want
to allow that response to run to completion.