refactor: Update Concourse build to only package lib module
For publishing to PyPI we only care about providing the contents of the `lib` module in
the `Concourse` package because that is where the reusable portions of the code are
located. This updates the BUILD metadata to only include that module in the generated
package and moves the `pipelines` module into a separate distribution for packaging and
distributing as a secondary package which will depend on the `ol-concourse` package for
proper operation.