chore: Add permissions to MITx Online course exports in Open
The MIT Open application processes the course XML exports that are generated by the Open
edX extracts daily data pipeline so that it can populate a search index of course assets
that is displayed to end users as a content discovery capability in the app. This adds
permission to access the MITx Online course exports.