fix: Add permissions for Airbyte to interact with AWS Lake Formation
We are using the `destination-aws-datalake` connector for managing tables in our lake,
which requires granting access to the Airbyte instance for those APIs. An artifact of
how the plugin is written also requires explicitly setting a role to assume with STS, so
this creates a separate role for assuming via that path. In the future we should remove
that requirement in the plugin so that it will natively use the existing instance
profile for authz on those APIs.