hack: Add route table entry for Micromasters to data VPC
The Micromasters RDS instance is in its own VPC for historical reasons, and because the application
is being deprecated there is no reason to do the work to migrate to a more up to date
setup. This just adds the route table entry needed to ensure that the communication
isn't lost on network stack updates.