feat: Add helper functions for infrastructure pipelines
We have established a repeatable and reusable pattern of building EC2 AMIs with Packer
and then deploying them in stages with Pulumi. Writing pipelines for that process has
largely been copy/paste which makes it cumbersome. This adds a set of function
definitions and some supporting models to reduce the amount of information/work that is
required to create instances of those pipelines. This also adds a full pipeline for
deploying Airbyte when the upstream project releases new versions.