fix: Upgrade Airbyte vault mount to KV V2
The Airbyte application appears to require a V2 mount for storage of its secrets. The
path that it writes to uses the `data/` prefix and then when trying to execute it is
unable to retrieve the secret that it writes which indicates that it is trying to read
from the path in a manner consistent with how the v2 mount operates. This updates the
templates to use the secret access format needed for a v2 mount and updates the Pulumi
code to create a v2 mount type. It also moves the secrets needed for the server to
operate into SOPS files and populates them in Vault automatically.