feat: Add pipeline for ocw_site deployment
This adds a pipeline definition for the ocw_site Pulumi project. As part of that this
addresses a behavioral aspect of Pydantic where validators only run on attribute
assignment, not attribute mutation. This meant that duplicate resources and
resource_types could be created when the attribute was mutated, so a fix is added in the
pulumi_jobs_chain to assign on each iteration to trigger the validators. As part of
running in the context of a Concourse pipeline the Fastly API key is needed, so this
also addresses that need by using an explicity Fastly provider and aliases it to avoid
triggering replacement of child resources.