fix: Resolve triggering of first Pulumi stage
The AMI jobs being passed as a dependency to the Pulumi step wasn't triggering because
of their being mutated during the looping over subsequent pipeline stages. This applies
a fix that passes copies of the dependencies into the nested function call so that the
removal of the trigger for subsequent steps isn't reflected in the earlier stage.