refactor: Align data warehouses around data lifecycle
The warehouse definitions and S3 buckets in the data lake make more sense to be modeled
around the stages of the data lifecycle than to arbitrarily segment them by business
units. The original motivation was for cost allocation purposes but it introduces
unnecessary complexity. This _adds_ the new definitions so that we can migrate to them
first, and a subsequent commit will _remove_ the old ones.