fix: Update lake database names to work better with dbt
By default dbt will append a custom schema name to the end of the existing schema
name. By moving the `data_stage` to the end, it makes it easier to make it materialize
the different model stages to the proper database name without having to do a bunch of
custom configuration. https://docs.getdbt.com/docs/building-a-dbt-project/building-models/using-custom-schemas