feat: Add a PyInfra component for the Traefik proxy
We have a small mix of proxy services that we depend on, with Caddy being the newest
addition. Unfortunately Caddy has proven to be fragile in some cases, and debugging is
challenging due to the multiple supported configuration syntaxes that are not all fully
documented. This adds a component to allow us to use Traefik in its place (particularly
the automatic TLS management).