fix: Use app role instead of admin for Airbyte DB connection
The app role definition automatically coalesces to use a defined role so that ownership
of tables/columns/etc. is shared regardless of which instance/lease are being used. With
the admin role that is not handled automatically so it leads to ownership/permissions
issues as the credentials are cycled and migrations get executed.