refactor: Use version info module for edX in MFE pipeline
The MFE pipelines are currently set to use a single version across all environments
which is problematic during our upgrade cycles. This refactors the structure to use a
single pipeline object for all MFEs used for a given deployment segmented by
release. This means that all MITx MFEs running the Olive release are combined in a
single pipeline, etc.

In order to ensure that the right versions are used in the right places this refactor
also brings in the canonical `version_matrix` module from the `bridge` package. There is
also some refactoring of the `version_matrix` into separate modules within the
`settings.openedx` namespace that makes it easier to read and reason about.