fix: Configure Postgres replicas to allow for data integration
Postgres read replicas will cancel long-running queries after 30 seconds if the data
being replicated from the primary instance modifies the table being accessed. This
causes synchronization failures in Airbyte due to the need to copy entire tables which
will typically require more than 30 seconds which causes syncs to fail on applications
that are being actively used. This updates the parameter group to allow the read
replica to provide feedback to the primary when there are active queries so that it
doesn't terminate those requests.