config: Enable library authoring MFE and Blockstore for Olive
The blockstore application is the forward-looking solution to composable course content,
as well as a building block for exposing individual learning resources to external
systems such as Canvas. This enables the library authoring MFE which is used for
creating those components, and toggles the flags to enable the blockstore feature.