fix: Downgrade requests due to failures in SAML integration
The 2.28.1 release changed some of the routines for how Requests validates TLS
certificates. This causes a failure in how the SAML integration operates that makes it
impossible to log in via Touchstone. This adds a forced dowgrade of requests to the
version used in Nutmeg until we can find a more permanent solution to the problem.