fix: Move edx release to .env file
Docker compose doesn't let you interpolate values in the compose file when they are
fetched through the `env_file` attribute. If there is a `.env` file in the context of
the compose file then it will be loaded automatically and early enough to use for
e.g. setting the tag to pull when the template runs.