fix: Don't automatically trigger deploys to production Pulumi stacks
Because we have pipeline templates that will filter which Pulumi stacks to execute for a
given pipeline based on e.g. edX versions for a given deployment there is the
possibility that a production deployment could get triggered automatically. Until we
have more automated checks in place we want to run trigger production Pulumi runs
manually. This adds logic to ensure that we don't accidentally end up with a pipeline
that has a production trigger based on meta pipelines and stack filters.