fix: Use SOPS for storing Concourse pipeline secrets
By placing the secret values in the stack config it contributes to code sprawl and
requires looking in multiple locations to figure out what secrets are being
maintained. This refactors the Concourse code to use SOPS for populating secrets in
Vault that are relied on for various pipelines. The reason this came up is that the
dockerhub credentials kept getting overwritten with invalid values after rotating the
password and forgetting to update the stack config.