fix: Remove extraneous 503 condition to fix robots.txt
The 503 response was not being assigned to anything, so it was erroneously attached to
the robots.txt response object. This is because the `ServiceVCLResponseObjects can only
be assigned a condition of type `REQUEST`, but the `ServiceVCLCondition` for 503 has a
type of `CACHE` and for some reason that was getting mapped to the robots.txt response
since the condition wasn't being consumed elsewhere.