version,refactor: Update pants and Python versions
Pants had been a constraint on what version of Python we could use in our containers due
to the need for it to have a compatible version to bootstrap from. It now contains its
own Python runtime so we can safely use whichever version we want for our actual project code.