feat: Update residential CI to Palm with how-to doc
This updates the version configuration to set the CI stage of the residential MITx
deployments to use the new Palm release. This also adds a corresponding `how-to`
document that explains how to do this in the future with the aim of making it easier for
anyone to manage the process.