config: Add bucket permissions for OCW Studio to sync videos
There are a number of videos in our legacy offline site mirror that we would like to
propagate to our new offline site configuration. This adds the IAM permissions to the
OCW Studio application to manage that process. Once we have moved all of the videos we
will likely want to remove these permissions.