fix: Prevent collision of Pulumi resources in Concourse pipelines
The identifier for the Pulumi resource provider was hard-coded, so if more than one
Pulumi chain is present in a given pipeline then the last one wins for the provider
definition. This updates the identifier to use the project name to avoid those
collisions. This also simplifies the Vault substructure pipeline logic a bit to use the
`combine_fragments` method.