fix: Publish our own image of hashicorp-release-resource
The upstream docker repository appears to have been removed, causing all of our
pipelines that rely on fetching the latest version of Hashicorp products to fail. This
adds  pipeline to build our own image from our fork of the repo and updates our resource
type definition to use that image.