feat,config: Create Vault role definition for Airbyte S3 sources
The S3 source connector for Airbyte currently (still) doesn't support authenticating via
the EC2 instance profile, instead requiring IAM keypairs. This adds a role definition
for Airbyte that will grant permission to all of the S3 sources that we need to read
from, simplifying the maintenance of permissions for those IAM pairs in the Airbyte setup.