config: Remove engine_version overrides for RDS component
The OLAmazonDB component resource config classes have a validator to ensure that we
don't specify engine versions that are not supported by AWS. Recently the versions we
were specifying became out of date so this logic started throwing errors.

Separately, we tell the component to ignore the engine_version after an instance has
been deployed due to configuration drift from the auto minor version upgrades that are
performed.

This removes custom version overrides for our RDS instances so that they all fall
through to the default values which were bumped in a separate commit.