config: Only parse the top level JSON for tracking logs in Vector
In the data platform we need to have a consistent type for the JSON attributes, but the
`event` field in particular flip-flops between string and object. Setting `max_depth`
field in the `parse_json` call in Vector renders objects as JSON-escaped strings

This also sets a destination for malformed messages in case we want to reprocess those fields.