feat: Add helper method for InstanceTypes enum
The instance types enum is intended as a means of managing evolution of instance types
deployed without having to find/replace across the whole repo. Sometimes we want to use
an instance type that is not commonly applied, so we have to route around the
`InstanceTypes` enum. This adds a `dereference` method to the enum to allow for using it
as a common pass-through, while also ensuring that we are using valid instance types.