config: Add dockerignore to exclude .git directory in edxapp image
The majority of the image size is due to the `COPY` step that is being done to load the
edx-platform repository into the image. The majority of that space is comprised of the
`.git` directory. This adds a dockerignore file to specify that that folder should be
excluded from the final product.