fix: Trigger workers.json creation during asset compilation for edxapp
The workers.json file that is needed for generating the appropriate JS bundle to run
proctoring is created as part of the plugin loading step for edx-proctoring. This adds a
step to run `collectstatic` before compiling the webpack assets in order to ensure that
the Django modules are loaded prior to JS compilation. This also explicitly sets the
static roots for LMS and studio.