config: Increase Traefik timeout for Airbyte backend
The Airbyte backend has some operations that can take a significant amount of time when
dealing with source connectors that have a lot of data or are slow to respond. This sets
the idle connection timeout in Traefik to 5 minutes to address 504 errors when
e.g. refreshing the schema for the GitHub source.