config,fix: Address GitHub rate limit issues with Concourse pipelines
Concourse pipelines that use the github release resource are often rate limited due to
accessing the API anonymously. As we continue to build and manage pipelines we will hit
this more frequently. This adds a default argument of the access token to allow for
authenticated requests with a higher rate limit.