config: Add MongoDB connection URI to Vault for Forum connection
Dagster is now fetching the Vault credentials from Vault directly. This means that we
need a way to retrieve the URI data from Vault rather than Consul. The username/password
are managed here by Pulumi, so this adds population of the URI string as well, fetched
from the MongoDB Atlas stack output.