config: Increase Vault cert validity to 2 years
The current setup limits the valid timeframe for the Vault certificate to 1 year, so
that at the turn of each new year the cert is expired and newly launched clusters
fail. This updates the cert validity to 2 years so that there is enough overlap for
cycling the clusters each new year.