feat,refactor: Initial stab at creating Leek config  through Pulumi
Leek will need to be able to speak to a large number of Celery queues. To manage that
config Pulumi is the logical focal point to pull it all together. We will generate the
full JSON document needed by Leek in Pulumi and populate it into Vault for doing a
simple retrieval into the `.env` file using Consul Template.