feat: Allow managing RDS DB versions by major only
In order to allow for keeping RDS databases up to date with patch releases we were using
te `auto_minor_version_upgrade`. That necessitated ignoring version changes, which makes
it impossible to cleanly upgrade versions. This adds logic for specifying only the major
version, while keeping the minor up to date. It also allows for pinning to a specific
full version.