fix: Address inconsistencies in parameter group for read replica
There are cases (specifically with MariaDB) where the parameter group family changes,
but the major engine version does not. This can lead to a situation where the read
replica needs to be upgraded first, and requires a new parameter group, but the primary
can not yet be upgraded and needs to hold onto the parameter group. This has a lot of
conditional logic for managing the case where the replica gets updated, and then the
primary will follow with a second application of the stack.