config: Allow RDS to automatically upgrade minor versions
The last refactoring allows the component to manage the DB version to unblock major
version upgrades. The logic as it exists now doesn't conflict with the auto minor
version upgrade. This way RDS can manage minor versions during maintenance windows, but
the component can also keep the DB up to date. The intent is so that the component
resource isn't the _only_ way that the minor version gets updated.