config: Move implicit path rewrites to the fetch stage for Learn
The MIT Learn frontend is managed by Fastly and stored in S3. The actual file to fetch
for most paths is the single `index.html`, but the site requests different paths due to
how React routing works. By rewriting the `req.url` we run the risk of accidentally
leaking that to the browser, leading to a 404. This moves the path rewriting to the
`fetch` stage using the `bereq.url` object.