refactor: Move postgres role creation statements inline
Having the role creation statements as a separate vault role definition causes confusion
and impedes our ability to cleanly set up new apps and environments without manual
intervention. The previous motivation for splitting the statements into separate blocks
was due to an unrelated issue with excessive readonly users being granted direct access
to tables and other objects. That is no longer the case so the role definitions can be
moved back in-line.