Install yarn (otherwise it tries to use the default 1.22 version)
This is a touch more involved. We actually need to enable Corepack, and then it will go get the appropriate yarn for us. Then, the publish command is different in yarn berry - so hopefully this is enough. (It will default to the "latest" tag and I think the version will come out of the package.json?)