fix: Use the Docker digest for pinning Keycloak version
The Docker image tag that is generated gets tied to the git hash of the ol-keycloak
repository, which isn't a stable identifier across builds. It is unlikely that specifier
will change as frequently as the built image. This leads to bugs with cached images
being used rather than the image that should be built into the AMI.