fix: Split tracking log locations for LMS and CMS
Because there are both LMS and CMS containers running via docker compose, they end up
competing for the ability to write to the on-disk location that Vector reads from. This
splits their usage of the physical disk to be namespaced by the service type. This also
updates the Vector config to watch _all_ of the tracking log files since some logs are
being written to a `.1` location simultaneously with logs being written to the primary file.