chore: Configure Renovate to maintain versions in versions.py
The versions.py file is a useful source of truth for what versions are
running. Unfortunately it frequently falls out of date as we pull versions dynamically
into the build pipelines. This adds information to the versions.py file that Renovate
can use to keep that file up to date as upstream versions change. https://docs.renovatebot.com/modules/manager/regex/