fix(traefik): configure trusted IPs for forwarded headers and fix field name
This fixes a typo in the field name (trusted_i_ps -> trusted_ips) and adds proper
configuration for trusted IP ranges in forwarded headers for both HTTP and HTTPS
entrypoints. This allows `X-Forwarded-*` headers to be passed along to Keycloak for
determining when to send CORS responses.

Also adds web origins configuration for the mitlearn client.