config: Don't store user in APISix session cookie
The session cookie set by APISix using the openid-connect plugin is substantially
large. This updates the configuration to _not_ store the user in the cookie in the hopes
of reducing the overall cookie size. This is related to 400 errors that are happening
for a subset of users due to the overall number and size of cookies stored on .mit.edu subdomains.