feat: Add startup, readiness, and liveness probes to K8s apps
We currently experience temporary 502s from the applications while they are doing a
deployment rollout. This adds startup, readiness, and liveness checks to the pods to
ensure that they are ready to serve traffic before they get incorporated into their
corresponding service. This depends on upstream changes in the applications in PRS:
- https://github.com/mitodl/mit-learn/pull/2247
- https://github.com/mitodl/mitxonline/pull/2652